Test: Make sure type section validation fails correctly
This commit adds tests and invalid wasm testfiles that checks the type section.

The file invalid_form.wasm is based on testsuite/address.wasm and was
edited to have the first entry of the type section include an invalid
form (0x60 => 0x61).

The file invalid_retcount.wasm is also based on testsuite/address.wasm
and was edited to have the first type section entry include two return
values instead of one.